prefs: extract reusable widgets to individual modules
Remove duplicate implementation from tab title dialog.
