+++ ddterm/app/tablabel.js
@@ -8,46 +8,8 @@ import Handy from 'gi://Handy';
 
 import Gettext from 'gettext';
 
-class EntryRow extends Handy.ActionRow {
-    static [GObject.GTypeName] = 'DDTermTabTitleEntryRow';
-
-    static [GObject.properties] = {
-        'text': GObject.ParamSpec.string(
-            'text',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            ''
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    #entry;
-
-    constructor(params) {
-        super(params);
-
-        this.#entry = new Gtk.Entry({
-            visible: true,
-            hexpand: true,
-            valign: Gtk.Align.CENTER,
-        });
-
-        this.bind_property(
-            'text',
-            this.#entry,
-            'text',
-            GObject.BindingFlags.SYNC_CREATE | GObject.BindingFlags.BIDIRECTIONAL
-        );
-
-        this.set_activatable(true);
-        this.set_activatable_widget(this.#entry);
-        this.add(this.#entry);
-    }
-}
+import { EntryRow } from '../pref/widgets/entryrow.js';
+import { SwitchRow } from '../pref/widgets/switchrow.js';
 
 export class TabTitleDialog extends Gtk.Dialog {
     static [GObject.GTypeName] = 'DDTermTabTitleDialog';
@@ -92,27 +54,32 @@ export class TabTitleDialog extends Gtk.Dialog {
             GObject.BindingFlags.BIDIRECTIONAL | GObject.BindingFlags.SYNC_CREATE
         );
 
-        const expander = new Handy.ExpanderRow({
+        const toggle = new SwitchRow({
             visible: true,
-            show_enable_switch: true,
             use_underline: true,
             title: Gettext.gettext('Use Custom Tab Title'),
         });
 
-        expander.add(entry);
+        this.bind_property(
+            'use-custom-title',
+            toggle,
+            'active',
+            GObject.BindingFlags.BIDIRECTIONAL | GObject.BindingFlags.SYNC_CREATE
+        );
 
         this.bind_property(
             'use-custom-title',
-            expander,
-            'enable-expansion',
-            GObject.BindingFlags.BIDIRECTIONAL | GObject.BindingFlags.SYNC_CREATE
+            entry,
+            'sensitive',
+            GObject.BindingFlags.SYNC_CREATE
         );
 
         const group = new Handy.PreferencesGroup({
             visible: true,
         });
 
-        group.add(expander);
+        group.add(toggle);
+        group.add(entry);
 
         this.get_content_area().add(group);
     }
