+++ ddterm/pref/util.js
@@ -2,19 +2,19 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import GLib from 'gi://GLib';
 import GObject from 'gi://GObject';
 import Gio from 'gi://Gio';
 import Gtk from 'gi://Gtk';
 
 import Gi from 'gi';
 
-export const AdwOrHdy = Gi.require(Gtk.get_major_version() === 3 ? 'Handy' : 'Adw');
+import { ComboTextRow, ComboTextItem } from './widgets/comborow.js';
+import { SwitchRow } from './widgets/switchrow.js';
 
-export const {
-    PreferencesRow,
-    ActionRow,
-} = AdwOrHdy;
+export { ActionRow } from './widgets/actionrow.js';
+
+export const AdwOrHdy = Gi.require(Gtk.get_major_version() === 3 ? 'Handy' : 'Adw');
+export const { PreferencesRow } = AdwOrHdy;
 
 export function add_reset_button(row, settings, key, gettext_domain) {
     const button = Gtk.Button.new_from_icon_name.length === 1
@@ -40,12 +40,7 @@ export function add_reset_button(row, settings, key, gettext_domain) {
             : Gtk.RevealerTransitionType.SLIDE_RIGHT,
     });
 
-    if (row.add_suffix)
-        row.add_suffix(container);
-    else if (row.add_action) // For Handy.ExpanderRow - otherwise gets added as a child row
-        row.add_action(container);
-    else
-        row.add(container);
+    row.add_suffix(container);
 
     settings.bind_writable(key, button, 'sensitive', false);
 
@@ -70,72 +65,6 @@ export function add_reset_button(row, settings, key, gettext_domain) {
     return button;
 }
 
-export const SwitchRow = AdwOrHdy.SwitchRow ?? class extends ActionRow {
-    static [GObject.GTypeName] = 'DDTermSwitchRow';
-
-    static [GObject.properties] = {
-        'active': GObject.ParamSpec.boolean(
-            'active',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            false
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    #slider;
-    #slider_binding;
-
-    constructor(params) {
-        super(params);
-
-        this.#slider = new Gtk.Switch({
-            valign: Gtk.Align.CENTER,
-            can_focus: false,
-            visible: true,
-        });
-
-        this.set_activatable(true);
-        this.set_activatable_widget(this.#slider);
-        this.add(this.#slider);
-
-        this.connect('notify::action-name', this.#update_active_binding.bind(this));
-        this.#update_active_binding();
-
-        this.bind_property(
-            'action-name',
-            this.#slider,
-            'action-name',
-            GObject.BindingFlags.SYNC_CREATE
-        );
-
-        this.bind_property(
-            'action-target',
-            this.#slider,
-            'action-target',
-            GObject.BindingFlags.SYNC_CREATE
-        );
-    }
-
-    #update_active_binding() {
-        if (this.action_name) {
-            this.#slider_binding?.unbind();
-            this.#slider_binding = null;
-        } else if (!this.#slider_binding) {
-            this.#slider_binding = this.bind_property(
-                'active',
-                this.#slider,
-                'active',
-                GObject.BindingFlags.BIDIRECTIONAL | GObject.BindingFlags.SYNC_CREATE
-            );
-        }
-    }
-};
-
 function create_switch_row({
     settings,
     key,
@@ -155,469 +84,33 @@ function create_switch_row({
     return row;
 }
 
-export const EntryRow = AdwOrHdy.EntryRow ?? class extends ActionRow {
-    static [GObject.GTypeName] = 'DDTermEntryRow';
+function create_combo_text_row({
+    settings,
+    key,
+    model,
+    flags = Gio.SettingsBindFlags.DEFAULT,
+    gettext_domain,
+    ...params
+}) {
+    if (model && !(model instanceof GObject.Object))
+        model = ComboTextItem.create_list(model);
 
-    static [GObject.properties] = {
-        'text': GObject.ParamSpec.string(
-            'text',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            ''
-        ),
-    };
+    const row = new ComboTextRow({
+        visible: true,
+        use_underline: true,
+        ...params,
+    });
 
-    static {
-        GObject.registerClass(this);
-    }
+    row.bind_name_model(model);
 
-    #entry;
+    settings.bind(key, row, 'value', flags);
+    add_reset_button(row, settings, key, gettext_domain);
 
-    constructor(params) {
-        super(params);
-
-        this.#entry = new Gtk.Entry({
-            visible: true,
-            hexpand: true,
-            valign: Gtk.Align.CENTER,
-        });
-
-        this.bind_property(
-            'text',
-            this.#entry,
-            'text',
-            GObject.BindingFlags.SYNC_CREATE | GObject.BindingFlags.BIDIRECTIONAL
-        );
-
-        this.set_activatable(true);
-        this.set_activatable_widget(this.#entry);
-
-        if (this.add_suffix)
-            this.add_suffix(this.#entry);
-        else
-            this.add(this.#entry);
-    }
-};
-
-export const SpinRow = AdwOrHdy.SpinRow ?? class extends ActionRow {
-    static [GObject.GTypeName] = 'DDTermSpinRow';
-
-    static [GObject.properties] = {
-        'adjustment': GObject.ParamSpec.object(
-            'adjustment',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            Gtk.Adjustment
-        ),
-        'digits': GObject.ParamSpec.int(
-            'digits',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            0,
-            20,
-            0
-        ),
-        'numeric': GObject.ParamSpec.boolean(
-            'numeric',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            false
-        ),
-        'snap-to-ticks': GObject.ParamSpec.boolean(
-            'snap-to-ticks',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            false
-        ),
-        'value': GObject.ParamSpec.double(
-            'value',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            -Number.MAX_VALUE,
-            Number.MAX_VALUE,
-            0
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    #spin;
-
-    constructor(params) {
-        super(params);
-
-        this.#spin = new Gtk.SpinButton({
-            visible: true,
-            hexpand: true,
-            valign: Gtk.Align.CENTER,
-        });
-
-        for (const prop of ['snap-to-ticks', 'numeric', 'digits', 'adjustment'])
-            this.bind_property(prop, this.#spin, prop, GObject.BindingFlags.SYNC_CREATE);
-
-        this.bind_property(
-            'value',
-            this.#spin,
-            'value',
-            GObject.BindingFlags.SYNC_CREATE | GObject.BindingFlags.BIDIRECTIONAL
-        );
-
-        this.set_activatable(true);
-        this.set_activatable_widget(this.#spin);
-
-        if (this.add_suffix)
-            this.add_suffix(this.#spin);
-        else
-            this.add(this.#spin);
-    }
-};
-
-export const INVALID_LIST_POSITION = Gtk.INVALID_LIST_POSITION ?? GLib.MAXUINT32;
-
-export const ComboRow = AdwOrHdy.ComboRow.prototype.bind_name_model
-    ? class extends AdwOrHdy.ComboRow {
-        static [GObject.GTypeName] = 'DDTermComboRow';
-
-        static [GObject.properties] = {
-            'selected': GObject.ParamSpec.uint(
-                'selected',
-                null,
-                null,
-                GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-                0,
-                GLib.MAXUINT32,
-                INVALID_LIST_POSITION
-            ),
-            'selected-item': GObject.ParamSpec.object(
-                'selected-item',
-                null,
-                null,
-                GObject.ParamFlags.READABLE,
-                GObject.Object
-            ),
-        };
-
-        static {
-            GObject.registerClass(this);
-        }
-
-        constructor(params) {
-            super(params);
-
-            this.connect('notify::selected-index', () => {
-                this.notify('selected');
-                this.notify('selected-item');
-            });
-        }
-
-        get selected() {
-            if (this.selected_index < 0)
-                return INVALID_LIST_POSITION;
-
-            return this.selected_index;
-        }
-
-        set selected(index) {
-            if (index === INVALID_LIST_POSITION)
-                this.selected_index = -1;
-            else
-                this.selected_index = index;
-        }
-
-        get selected_item() {
-            const { selected } = this;
-            const model = this.get_model();
-
-            if (!model || selected >= model.get_n_items())
-                return null;
-
-            return model.get_item(selected);
-        }
-    } : class extends AdwOrHdy.ComboRow {
-        static [GObject.GTypeName] = 'DDTermComboRow';
-
-        static {
-            GObject.registerClass(this);
-        }
-
-        bind_name_model(model, get_name_func) {
-            this.set_get_name_func(get_name_func);
-            this.set_model(model);
-        }
-
-        set_get_name_func(get_name_func) {
-            this.set_expression(Gtk.ClosureExpression.new(String, get_name_func, null));
-        }
-    };
-
-export class ComboTextItem extends GObject.Object {
-    static [GObject.GTypeName] = 'DDTermComboTextItem';
-
-    static [GObject.properties] = {
-        'name': GObject.ParamSpec.string(
-            'name',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.CONSTRUCT_ONLY,
-            null
-        ),
-        'value': GObject.ParamSpec.string(
-            'value',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.CONSTRUCT_ONLY,
-            null
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    static create_list(mapping) {
-        const model = Gio.ListStore.new(this);
-
-        for (const [value, name] of Object.entries(mapping))
-            model.append(new this({ value, name }));
-
-        return model;
-    }
+    return row;
 }
 
-export class ComboTextRow extends ComboRow {
-    static [GObject.GTypeName] = 'DDTermComboTextRow';
-
-    static [GObject.properties] = {
-        'value': GObject.ParamSpec.string(
-            'value',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            null
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    constructor(params) {
-        super(params);
-
-        this.connect('notify::selected-item', () => this.notify('value'));
-    }
-
-    bind_name_model(model) {
-        super.bind_name_model(model, v => v.name);
-    }
-
-    get value() {
-        return this.selected_item?.value ?? null;
-    }
-
-    set value(new_value) {
-        const model = this.get_model();
-
-        if (!model)
-            return;
-
-        for (let i = 0; i < model.get_n_items(); i++) {
-            const item = model.get_item(i);
-
-            if (item.value === new_value) {
-                this.selected = i;
-                return;
-            }
-        }
-
-        this.selected = INVALID_LIST_POSITION;
-    }
-
-    static create({
-        settings,
-        key,
-        model,
-        flags = Gio.SettingsBindFlags.DEFAULT,
-        gettext_domain,
-        ...params
-    }) {
-        if (model && !(model instanceof GObject.Object))
-            model = ComboTextItem.create_list(model);
-
-        const row = new this({
-            visible: true,
-            use_underline: true,
-            ...params,
-        });
-
-        row.bind_name_model(model);
-
-        settings.bind(key, row, 'value', flags);
-        add_reset_button(row, settings, key, gettext_domain);
-
-        return row;
-    }
-}
-
-export const StringObject = Gtk.StringObject ?? class extends GObject.Object {
-    static [GObject.GTypeName] = 'DDTermStringObject';
-
-    static [GObject.properties] = {
-        'string': GObject.ParamSpec.string(
-            'string',
-            null,
-            null,
-            GObject.ParamFlags.READABLE,
-            null
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    #string = null;
-
-    get string() {
-        return this.#string;
-    }
-
-    get_string() {
-        return this.#string;
-    }
-
-    static new(s) {
-        const o = new this();
-
-        o.#string = s;
-
-        return o;
-    }
-};
-
-export const StringList = Gtk.StringList ?? class extends Gio.ListStore {
-    static [GObject.GTypeName] = 'DDTermStringList';
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    constructor(params) {
-        super({
-            item_type: StringObject,
-            ...params,
-        });
-    }
-
-    static new(strings) {
-        const o = new this();
-
-        o.splice(0, 0, strings);
-
-        return o;
-    }
-
-    append(s) {
-        super.append(StringObject.new(s));
-    }
-
-    find(s) {
-        const n = this.get_n_items();
-
-        for (let i = 0; i < n; i++) {
-            if (this.get_string(i) === s)
-                return i;
-        }
-
-        return INVALID_LIST_POSITION;
-    }
-
-    get_string(position) {
-        return this.get_item(position).string;
-    }
-
-    splice(position, n_removals, additions) {
-        super.splice(position, n_removals, additions.map(v => StringObject.new(v)));
-    }
-};
-
-export class ScaleRow extends ActionRow {
-    static [GObject.GTypeName] = 'DDTermScaleRow';
-
-    static [GObject.properties] = {
-        'adjustment': GObject.ParamSpec.object(
-            'adjustment',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            Gtk.Adjustment
-        ),
-        'round-digits': GObject.ParamSpec.int(
-            'round-digits',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            -1,
-            GLib.MAXINT32,
-            -1
-        ),
-        'digits': GObject.ParamSpec.int(
-            'digits',
-            null,
-            null,
-            GObject.ParamFlags.READWRITE | GObject.ParamFlags.EXPLICIT_NOTIFY,
-            -1,
-            32,
-            1
-        ),
-    };
-
-    static {
-        GObject.registerClass(this);
-    }
-
-    #scale;
-
-    constructor(params) {
-        super(params);
-
-        this.#scale = new Gtk.Scale({
-            visible: true,
-            hexpand: true,
-            valign: Gtk.Align.CENTER,
-            draw_value: true,
-        });
-
-        for (const prop of ['round-digits', 'digits', 'adjustment'])
-            this.bind_property(prop, this.#scale, prop, GObject.BindingFlags.SYNC_CREATE);
-
-        this.set_activatable(true);
-        this.set_activatable_widget(this.#scale);
-
-        if (this.add_suffix)
-            this.add_suffix(this.#scale);
-        else
-            this.add(this.#scale);
-    }
-
-    set_format_value_func(formatter) {
-        if (this.#scale.set_format_value_func)
-            this.#scale.set_format_value_func(formatter);
-        else
-            this.#scale.connect('format-value', formatter);
-    }
-}
-
-export class ExpanderRow extends AdwOrHdy.ExpanderRow {
-    static [GObject.GTypeName] = 'DDTermExpanderRow';
+class ExpanderRow extends AdwOrHdy.ExpanderRow {
+    static [GObject.GTypeName] = 'DDTermExpanderRowEx';
 
     static [GObject.properties] = {
         'settings': GObject.ParamSpec.object(
@@ -652,7 +145,7 @@ export class ExpanderRow extends AdwOrHdy.ExpanderRow {
     }
 
     add_combo_text_row(params) {
-        const row = ComboTextRow.create({
+        const row = create_combo_text_row({
             settings: this.settings,
             gettext_domain: this.gettext_domain,
             ...params,
@@ -668,6 +161,13 @@ export class ExpanderRow extends AdwOrHdy.ExpanderRow {
             super.add_row(row);
         else
             super.add(row);
+    }
+
+    add_suffix(widget) {
+        if (super.add_suffix)
+            super.add_suffix(widget);
+        else
+            super.add_action(widget);
     }
 
     static create({
@@ -742,7 +242,7 @@ export class PreferencesGroup extends AdwOrHdy.PreferencesGroup {
     }
 
     add_combo_text_row(params) {
-        const row = ComboTextRow.create({
+        const row = create_combo_text_row({
             settings: this.settings,
             gettext_domain: this.gettext_domain,
             ...params,
